refact: move types declared in utils/ to types/
