+++ apps/mobile/app/(authenticated)/(tabs)/(signer,explorer,converter)/signer/nostr/account/[npub]/note.tsx
@@ -55,13 +55,15 @@ import { usePriceStore } from '@/store/price'
 import { useSettingsStore } from '@/store/settings'
 import { useZapFlowStore } from '@/store/zapFlow'
 import { Colors } from '@/styles'
-import { type NostrKind0Profile } from '@/types/models/Nostr'
+import type {
+  NostrKind0Profile,
+  DecodedNostrContent,
+  FetchedNoteData
+} from '@/types/models/Nostr'
 import { type PaymentMethod } from '@/types/models/PaymentMethod'
 import { formatNostrCardDate } from '@/utils/format'
 import { getPubKeyHexFromNpub, validateNip05 } from '@/utils/nostr'
 import {
-  type DecodedNostrContent,
-  type FetchedNoteData,
   decodeNostrContent,
   extractEnhancedZapTags,
   extractPubpayTags,
