+++ apps/mobile/utils/nostrNoteThread.ts
@@ -1,4 +1,4 @@
-const EVENT_ID_HEX = /^[0-9a-fA-F]{64}$/
+import { NOSTR_EVENT_ID_HEX } from '@/constants/nostr'
 
 function listValidETags(tags: string[][]): string[][] {
   return tags.filter(
@@ -6,7 +6,7 @@ function listValidETags(tags: string[][]): string[][] {
       typeof t[0] === 'string' &&
       t[0].toLowerCase() === 'e' &&
       typeof t[1] === 'string' &&
-      EVENT_ID_HEX.test(t[1])
+      NOSTR_EVENT_ID_HEX.test(t[1])
   )
 }
